devSetup/demodata/requirements.txt
ebrains-kg-core
requests
marmotgraph==4.0.0